refactor: crawler code
